# api/openapi.yml
openapi: 3.0.3
info:
  title: Developer Utilities MCP Server
  description: |
    A comprehensive, high-performance utility server that consolidates essential development tools into a single service. 
    This server combines cryptography, data manipulation, text processing, identifier generation, time utilities, 
    and network tools to provide AI agents and developers with a one-stop solution for common development tasks.
    
    The server is designed to be secure, stateless, and highly performant, following enterprise-grade standards 
    while remaining accessible for open-source development.
  version: 1.0.0
  contact:
    name: Developer Utilities MCP Server
    url: https://github.com/example/dev-utilities
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: http://localhost:8080
    description: Local development server
  - url: https://api.example.com
    description: Production server
paths:
...
  /api/v1/transform/base64:
    post:
      tags:
        - Data Transformation
      summary: Base64 encode/decode
      description: Encode or decode data using Base64 (standard or URL-safe)
      operationId: transformBase64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Base64Request'
            examples:
              encode:
                summary: Base64 encode
                value:
                  content: "hello world"
                  action: "encode"
                  urlSafe: false
              decode:
                summary: Base64 decode
                value:
                  content: "aGVsbG8gd29ybGQ="
                  action: "decode"
                  urlSafe: false
      responses:
        '200':
          description: Base64 operation completed successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Base64Response'
        '400':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/transform/url:
    post:
      tags:
        - Data Transformation
      summary: URL encode/decode
      description: Encode or decode data using URL percent-encoding
      operationId: transformURL
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/URLEncodeRequest'
            examples:
              encode:
                summary: URL encode
                value:
                  content: "hello world & special chars"
                  action: "encode"
              decode:
                summary: URL decode
                value:
                  content: "hello%20world%20%26%20special%20chars"
                  action: "decode"
      responses:
        '200':
          description: URL operation completed successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/URLEncodeResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/transform/jwt/decode:
    post:
      tags:
        - Data Transformation
      summary: Decode JWT token
      description: Decode JWT token and display header and payload (without signature verification)
      operationId: decodeJWT
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JWTDecodeRequest'
            examples:
              basic:
                summary: JWT decode
                value:
                  token: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c"
      responses:
        '200':
          description: JWT decoded successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/JWTDecodeResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/transform/compress:
    post:
      tags:
        - Data Transformation
      summary: Compress/decompress data
      description: Compress or decompress data using Gzip or Zlib with Base64 input/output
      operationId: compressData
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CompressionRequest'
            examples:
              compress:
                summary: Compress with Gzip
                value:
                  content: "This is some text that will be compressed"
                  action: "compress"
                  algorithm: "gzip"
      responses:
        '200':
          description: Compression operation completed successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/CompressionResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalError'

  # Identifier Generation Endpoints
...
  /api/v1/time/convert:
    post:
      tags:
        - Time Utilities
      summary: Convert time formats
      description: Convert between Unix timestamps, ISO 8601, and human-readable formats
      operationId: convertTime
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TimeConvertRequest'
            examples:
              unix_to_iso:
                summary: Unix timestamp to ISO 8601
                value:
                  input: "1640995200"
                  inputFormat: "unix"
                  outputFormat: "iso8601"
              iso_to_human:
                summary: ISO 8601 to human readable
                value:
                  input: "2022-01-01T00:00:00Z"
                  inputFormat: "iso8601"
                  outputFormat: "human"
      responses:
        '200':
          description: Time conversion completed successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/TimeConvertResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      required:
        - success
        - data
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          description: Response data (varies by endpoint)
    ErrorResponse:
      type: object
      required:
        - success
        - error
      properties:
        success:
          type: boolean
          example: false
        error:
          $ref: '#/components/schemas/APIError'
    APIError:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          description: Error code for programmatic handling
          enum:
            - VALIDATION_ERROR
            - AUTHENTICATION_ERROR
            - AUTHORIZATION_ERROR
            - RATE_LIMIT_ERROR
            - INTERNAL_ERROR
            - SERVICE_UNAVAILABLE
            - NOT_FOUND
            - METHOD_NOT_ALLOWED
          example: "VALIDATION_ERROR"
        message:
          type: string
          description: Human-readable error message
          example: "Invalid request format"
        details:
          type: string
          description: Additional error details
          example: "Field 'algorithm' is required"

    # Cryptography Schemas
...
    Base64Request:
      type: object
      required:
        - content
        - action
      properties:
        content:
          type: string
          description: Content to encode/decode
          example: "hello world"
        action:
          type: string
          description: Operation to perform
          enum: [encode, decode]
          example: "encode"
        urlSafe:
          type: boolean
          description: Use URL-safe Base64 variant
          default: false
          example: false
    Base64Response:
      type: object
      required:
        - result
      properties:
        result:
          type: string
          description: Encoded/decoded result
          example: "aGVsbG8gd29ybGQ="
    URLEncodeRequest:
      type: object
      required:
        - content
        - action
      properties:
        content:
          type: string
          description: Content to encode/decode
          example: "hello world & special chars"
        action:
          type: string
          description: Operation to perform
          enum: [encode, decode]
          example: "encode"
    URLEncodeResponse:
      type: object
      required:
        - result
      properties:
        result:
          type: string
          description: Encoded/decoded result
          example: "hello%20world%20%26%20special%20chars"
    JWTDecodeRequest:
      type: object
      required:
        - token
      properties:
        token:
          type: string
          description: JWT token to decode
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c"
    JWTDecodeResponse:
      type: object
      required:
        - header
        - payload
        - signatureVerified
      properties:
        header:
          type: object
          description: JWT header
          example:
            alg: "HS256"
            typ: "JWT"
        payload:
          type: object
          description: JWT payload
          example:
            sub: "1234567890"
            name: "John Doe"
            iat: 1516239022
        signatureVerified:
          type: boolean
          description: Signature verification status (always false)
          example: false
    CompressionRequest:
      type: object
      required:
        - content
        - action
        - algorithm
      properties:
        content:
          type: string
          description: Content to compress/decompress (Base64 for decompress)
          example: "This is some text that will be compressed"
        action:
          type: string
          description: Operation to perform
          enum: [compress, decompress]
          example: "compress"
        algorithm:
          type: string
          description: Compression algorithm
          enum: [gzip, zlib]
          example: "gzip"
    CompressionResponse:
      type: object
      required:
        - result
      properties:
        result:
          type: string
          description: Compressed/decompressed result (Base64 for compress)
          example: "H4sIAAAAAAAA/..."

    # Identifier Generation Schemas
...
    TimeConvertRequest:
      type: object
      required:
        - input
        - inputFormat
        - outputFormat
      properties:
        input:
          type: string
          description: Input time value
          example: "1640995200"
        inputFormat:
          type: string
          description: Input format
          enum: [unix, unix_ms, iso8601, rfc3339, human]
          example: "unix"
        outputFormat:
          type: string
          description: Output format
          enum: [unix, unix_ms, iso8601, rfc3339, human]
          example: "iso8601"
    TimeConvertResponse:
      type: object
      required:
        - result
      properties:
        result:
          type: string
          description: Converted time
          example: "2022-01-01T00:00:00Z"
...
  responses:
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            missing_field:
              summary: Missing required field
              value:
                success: false
                error:
                  code: "VALIDATION_ERROR"
                  message: "Invalid request format"
                  details: "Field 'algorithm' is required"
            invalid_value:
              summary: Invalid field value
              value:
                success: false
                error:
                  code: "VALIDATION_ERROR"
                  message: "Invalid request format"
                  details: "Field 'algorithm' must be one of: md5, sha1, sha256, sha512"
...
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            generic:
              summary: Generic internal error
              value:
                success: false
                error:
                  code: "INTERNAL_ERROR"
                  message: "Internal server error"
                  details: "An unexpected error occurred"
...
security:
  - ApiKeyAuth: []
  - BearerAuth: []
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for authentication
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for authentication